requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
certifi
redis
openai
kiwipiepy
pillow
pytesseract
PyPDF2
python-multipart
faiss-cpu


#keybert
#scikit-learn
#sentence-transformers
#numpy